chore(refactor): split SessionInspector.tsx into a SessionInspector/ directory
Final audit item. Same pattern as the Home/ split: 14 sub-components +
helpers + shared types extracted into siblings. None of the
sub-components reached into SessionInspector's outer scope beyond
their props.

Layout:
  web/src/components/SessionInspector/
    index.tsx              the SessionInspector orchestrator
    types.ts               Tab / RoleCategory
    utils.ts               formatTs / formatRel / categoryOf /
                           loadVisibleCategories / saveVisibleCategories
                           + TOOL_ONLY_RE / PAGE_SIZE / ALL_CATEGORIES /
                           DEFAULT_VISIBLE
    Header.tsx             (with PIP_CLASS / STATE_LABEL colocated)
    Banner.tsx
    Tabs.tsx
    SessionActions.tsx     (with shellQuote colocated — single caller)
    TranscriptBody.tsx     (largest leaf — find bar + tab gating)
    TranscriptSearchBar.tsx
    TranscriptFilter.tsx
    Transcript.tsx
    Turn.tsx
    ToolTurn.tsx           (with oneLineSummary / capRaw / RAW_CAP_BYTES
                            colocated — single caller)
    Highlighted.tsx
    MemoryTab.tsx
    MemorySection.tsx
    MemoryRow.tsx
    Artefacts.tsx          (with dedupeTouches colocated — single caller)

External import path (`Home → "./SessionInspector"`) still resolves —
TypeScript walks to index.tsx.

LOC: index.tsx 1292 → 292 lines (-77%, the orchestrator only). Sibling
files total ~1066 lines + 4 lines shared types + 61 lines utils. Net
+66 lines across the directory in exchange for navigable boundaries.

Helpers that only have one caller (shellQuote, oneLineSummary, capRaw,
dedupeTouches) live alongside that caller rather than in utils — no
forced sharing for one-shot helpers.

No behavioural change.
