+++ web/src/components/SessionInspector/types.ts
@@ -1,0 +1,4 @@
+// Shared SessionInspector types. Extracted from SessionInspector/index.tsx
+// so sibling components can import them directly.
+export type Tab = "transcript" | "artefacts" | "memory";
+export type RoleCategory = "user" | "assistant" | "tools" | "system" | "thinking";
